update database naming
With this all of the sriov related plugin files are now consistently
prepended with sriov- in the /tmp/ directory.

Now the file being used by the plugin is unique, so it should not
conflict in the event we are using multiple of these similar CNM
plugins (ie, VPP, OVSDPDK, SRIOV).

Signed-off-by: Eric Ernst <eric.ernst@intel.com>